Fixed variable types for Task 7

+++ activity/Task7.java
@@ -8,9 +8,9 @@ public class Task7 {
     public static void main(String[] args) {
         Scanner scn = new Scanner(System.in);
         System.out.print("ENTER FIRST NUMBER: ");
-        double num1 = scn.nextInt();
+        int num1 = scn.nextInt();
         System.out.print("ENTER SECOND NUMBER: ");
-        double num2 = scn.nextInt();
+        int num2 = scn.nextInt();
         scn.close();
 
         Task7 process = new Task7();
@@ -20,15 +20,15 @@ public class Task7 {
         System.out.println("Divide: " + process.divideTwoNumbers(num1, num2));
     }
 
-    double addTwoNumbers(double a, double b){
+    int addTwoNumbers(int a, int b){
         return a + b;
     }
 
-    double subtractTwoNumbers(double a, double b){
+    int subtractTwoNumbers(int a, int b){
         return a - b;
     }
 
-    double multiplyTwoNumbers(double a, double b){
+    int multiplyTwoNumbers(int a, int b){
         return a * b;
     }
